requirements.txt
# Core functionality
streamlit==1.28.1
streamlit-autorefresh==0.0.1
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0

# Visualization
plotly==5.17.0
altair==5.1.2

# PostgreSQL database adapter
psycopg2-binary==2.9.5

# Apache Spark
pyspark==3.3.2
findspark==2.0.1
py4j==0.10.9.7

# Kafka and fake data generation
kafka-python==2.0.2
faker==19.6.2


sqlalchemy==2.0.30

